Revert "update snapshots"
This reverts commit bbf1e82d8e3eb6bc1f17cf1844d0e04363532042.
