+++ client-generator/src/eteTest/java/com/fern/java/client/cli/ClientGeneratorEteTest.java
@@ -47,4 +47,15 @@ public class ClientGeneratorEteTest {
         SnapshotTestRunner.snapshotGithub(
                 eteTestDirectory, expect, "java-client:latest", Optional.of(Map.of("unknown-as-optional", true)));
     }
+
+    @SnapshotName("local")
+    @Test
+    public void test_local() throws IOException {
+        Path currentPath = Paths.get("").toAbsolutePath();
+        Path eteTestDirectory = currentPath.endsWith("client-generator")
+                ? currentPath.resolve(Paths.get("src/eteTest"))
+                : currentPath.resolve(Paths.get("client-generator/src/eteTest"));
+        SnapshotTestRunner.snapshotLocalFiles(
+                eteTestDirectory, expect, "java-client:latest", Optional.of(Map.of("unknown-as-optional", true)));
+    }
 }
